% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastqc-plots.R
\name{plot_dup_stats}
\alias{plot_dup_stats}
\title{Plot FastQC total sequences stats}
\usage{
plot_dup_stats(..., type = c("ggplot2", "rbokeh"), group = NULL)
}
\description{
Plot FastQC total sequences stats
}

