% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc-plots.R
\name{plot_total_sequence_stats}
\alias{plot_total_sequence_stats}
\title{Plot FastQC total sequences stats}
\usage{
plot_total_sequence_stats(..., interactive = TRUE, geom = c("jitter",
  "point", "bar"))
}

