% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastqc-plots.R
\name{plot_sequence_quality}
\alias{plot_sequence_quality}
\title{Plot FastQC per base sequence quality}
\usage{
plot_sequence_quality(..., type = c("ggplot2", "rbokeh"))
}
\description{
Plot FastQC per base sequence quality
}

